회원가입 화면 분리, 프로젝트 이름 변경(패키지 경로에 영향)
